==> _previous/clustering/exploration/fédérale_2021/3-description_clusters.R <==
# Packages ---------------------------------------------------------------
library(dplyr)
library(ggplot2)
library(cluster)
library(factoextra)
library(tidyr)
library(ggcorrplot)

# Description des clusters -----------------------------------------------

#### Function
describe_clusters <- function(data, variables_to_describe, cluster_var){
  data$cluster_var <- data[[cluster_var]]
  ### variables to dummy
  non_numeric_vars <- variables_to_describe[!sapply(data[,variables_to_describe], is.numeric)]
  if (!purrr::is_empty(non_numeric_vars)) {
    df_dummy_only <- data |> 
      select(all_of(non_numeric_vars)) |> 
      fastDummies::dummy_columns(
        select_columns = non_numeric_vars,
        omit_colname_prefix = TRUE,
        remove_selected_columns = TRUE
      ) |> 
      janitor::clean_names()
    variables_to_describe <- c(variables_to_describe[!variables_to_describe %in% non_numeric_vars], names(df_dummy_only))
    df_description <- cbind(data, df_dummy_only) |> 
      tidyr::pivot_longer(
        cols = all_of(variables_to_describe),
        names_to = "variable",
        values_to = "value"
      )
  } else {
      df_description <- data |> 
        tidyr::pivot_longer(
          cols = all_of(variables_to_describe),
          names_to = "variable",
          values_to = "value"
        )
  }
  df_mean_all <- df_description |> 
    group_by(variable) |> 
    summarise(
      mean_value = mean(value),
      sd_value = sd(value)
    )
  df_mean_by_cluster <- df_description |> 
    group_by(cluster_var, variable) |> 
    summarise(mean_cluster = mean(value)) %>% 
    left_join(
      ., df_mean_all, by = "variable"
    ) |> 
    mutate(z_score = (mean_cluster - mean_value) / sd_value)
  return(df_mean_by_cluster)
}

variables_to_describe <- c(
  "educBHS",
#  "educHS",
  "educUniv",
  "incomeLow",
  "incomeMid",
  "incomeHigh",
  "ses_hetero",
  "ses_gai",
#  "ses_bisex",
#  "ses_sexOri_other",
  "immigrant",
  "male",
#  "female",
#  "ses_genderOther",
  "age34m",
  "age3554",
  "age55p",
  "langFr",
  "langEn",
  "ses_languageOther",
  "act_transport_Car",
#  "act_transport_SUV",
#  "act_transport_Moto",
  "act_transport_Walk",
#  "act_transport_Bicycle",
  "act_transport_PublicTransportation",
#  "act_Walk",
  "act_Gym",
  "act_TeamSport",
  "act_Run",
  "act_Yoga",
#  "act_Swimming",
#  "act_Other",
  "act_None",
  "act_Fishing",
  "act_Hunting",
  "act_VisitsMuseumsGaleries",
  "act_MotorizedOutdoorActivities",
#  "act_Volunteering",
  "cons_brand_MaR",
#  "cons_brand_OnlineOnly",
#  "cons_brand_BInd",
  "cons_brand_ChainesB",
  "cons_brand_GSurf",
  "cons_brand_Frip",
#  "cons_brand_Other",
  "cons_Meat",
  "cons_Vege",
  "cons_Vegan",
  "cons_coffee_TimH",
#  "cons_coffee_Other",
  "cons_coffee_Starbucks",
#  "cons_coffee_SC",
  "cons_coffee_McDo",
#  "cons_coffee_place_ind",
  "cons_coffee_place_noCoffee",
  "cons_Smoke",
#  "cons_SmokeStopping",
  "cons_SmokeStopped",
  "cons_SmokeNever",
#  "cons_VapeNation",
  "cons_noDrink",
  "cons_redWineDrink",
#  "cons_whiteWineDrink",
#  "cons_roseDrink",
#  "cons_sparklingDrink",
  "cons_regBeers",
  "cons_microBeers",
#  "cons_spiritDrink",
  "cons_cocktailsDrink",
#  "app_swag_Formel",
#  "app_swag_Classique",
#  "app_swag_Casual",
#  "app_swag_Sport",
#  "app_swag_Chic",
#  "app_swag_VintageHippBoheme",
#  "app_swag_Other",
#  "app_swag_Rock",
  "app_noTattoo",
  "ses_dwelling_app",
#  "ses_dwelling_loft",
  "ses_dwelling_condo",
#  "ses_dwelling_tour",
  "ses_dwelling_detachedHouse"#,
#  "ses_dwelling_townHouse",
#  "ses_dwelling_semiDetached",
#  "ses_dwelling_coop",
#  "ses_dwelling_HLM",
#  "ses_dwelling_mobile",
#  "ses_dwelling_other"#,
#  "ses_dwelling_house"
)


# Chemin du dossier où enregistrer les graphiques
output_dir <- "/home/alexab/Dropbox/Ulaval/CLESSN/datagotchi_federal_2024/_SharedFolder_datagotchi_federal_2024/clustering/graph"

# S'assurer que le dossier existe
if(!dir.exists(output_dir)){
  dir.create(output_dir, recursive = TRUE)
}

# Charger les bibliothèques nécessaires
library(dplyr)
library(ggplot2)
library(forcats)

# Description des clusters
df_mean_by_cluster <- describe_clusters(
  data_filtered,
  variables_to_describe = variables_to_describe,
  cluster_var = "cluster_6"
)

# Limiter les scores z entre -2 et 2 (optionnel)
df_mean_by_cluster <- df_mean_by_cluster %>%
  mutate(z_score_limited = ifelse(z_score > 2, 2, ifelse(z_score < -2, -2, z_score)))

df_mean_by_cluster <- df_mean_by_cluster %>%
  mutate(
    variable = case_when(
      variable == "educBHS" ~ "Secondaire",
      variable == "educUniv" ~ "Universitaire",
      variable == "incomeLow" ~ "Revenu faible",
      variable == "incomeMid" ~ "Revenu moyen",
      variable == "incomeHigh" ~ "Revenu élevé",
      variable == "ses_hetero" ~ "Hétéro",
      variable == "ses_gai" ~ "Gai",
      variable == "immigrant" ~ "Immigrant",
      variable == "male" ~ "Homme",
      variable == "age34m" ~ "Moins de 34 ans",
      variable == "age3554" ~ "35 à 54 ans",
      variable == "age55p" ~ "55 ans et plus",
      variable == "langFr" ~ "Français",
      variable == "langEn" ~ "Anglais",
      variable == "ses_languageOther" ~ "Autre langue",
      variable == "act_transport_Car" ~ "Voiture",
      variable == "act_transport_Walk" ~ "Marche",
      variable == "act_transport_PublicTransportation" ~ "Transport public",
      variable == "act_Gym" ~ "Gym",
      variable == "act_TeamSport" ~ "Sports d'équipe",
      variable == "act_Run" ~ "Course",
      variable == "act_Yoga" ~ "Yoga",
      variable == "act_None" ~ "Aucune activité",
      variable == "act_Fishing" ~ "Pêche",
      variable == "act_Hunting" ~ "Chasse",
      variable == "act_VisitsMuseumsGaleries" ~ "Visites de musées/galleries",
      variable == "act_MotorizedOutdoorActivities" ~ "Activités motorisées",
      variable == "cons_brand_MaR" ~ "Magasins à rayon ",
      variable == "cons_brand_ChainesB" ~ "Chaînes de boutique",
      variable == "cons_brand_GSurf" ~ "Grandes surface",
      variable == "cons_brand_Frip" ~ "Friperie",
      variable == "cons_Meat" ~ "Carnivore",
      variable == "cons_Vege" ~ "Végétarien",
      variable == "cons_Vegan" ~ "Végan",
      variable == "cons_coffee_TimH" ~ "Café Tim Hortons",
      variable == "cons_coffee_Starbucks" ~ "Café Starbucks",
      variable == "cons_coffee_McDo" ~ "Café McDonald's",
      variable == "cons_coffee_place_noCoffee" ~ "Pas de café",
      variable == "cons_Smoke" ~ "Fume",
      variable == "cons_SmokeStopped" ~ "A arrêté de fumer",
      variable == "cons_SmokeNever" ~ "N'a jamais fumé",
      variable == "cons_noDrink" ~ "Pas d'alcool",
      variable == "cons_redWineDrink" ~ "Vin rouge",
      variable == "cons_regBeers" ~ "Bières régulières",
      variable == "cons_microBeers" ~ "Microbrasseries",
      variable == "cons_cocktailsDrink" ~ "Cocktails",
      variable == "app_noTattoo" ~ "Pas de tatouages",
      variable == "ses_dwelling_app" ~ "Appartement",
      variable == "ses_dwelling_condo" ~ "Condo",
      variable == "ses_dwelling_detachedHouse" ~ "Maison individuelle",
      TRUE ~ "Autre"
    )
  )



# Obtenir la liste des clusters uniques
clusters <- unique(df_mean_by_cluster$cluster_var)

# Boucle sur chaque cluster
for(cluster_id in clusters){
  # Filtrer les données pour le cluster actuel
  df_cluster <- df_mean_by_cluster %>%
    filter(cluster_var == cluster_id)
  
  # Réordonner les variables par score z absolu
  df_cluster <- df_cluster %>%
    mutate(variable = fct_reorder(variable, abs(z_score), .desc = TRUE))
  
  # Inverser l'ordre des niveaux pour que les variables avec le score z absolu le plus élevé soient en haut
  df_cluster$variable <- fct_rev(df_cluster$variable)
  
# Ajouter une colonne pour la couleur en fonction du signe du z_score
df_cluster <- df_cluster %>%
  mutate(color = ifelse(z_score_limited >= 0, "#26aec6", "#eb1616"))


# Créer le graphique avec une ligne entre 0 et le point
graph <- ggplot(df_cluster, aes(x = z_score_limited, y = variable)) +
  # Lignes entre 0 et les points
  geom_segment(aes(x = 0, xend = z_score_limited, y = variable, yend = variable), 
               color = "grey", size = 1) +  # Épaisseur augmentée
  # Points avec couleur conditionnelle
  geom_point(aes(color = color), size = 3) +  # Points plus grands
  # Ligne verticale pour 0
  geom_vline(xintercept = 0, linetype = "dashed") +
  # Limites sur les axes
  coord_cartesian(xlim = c(-2, 2)) +
  # Personnalisation du thème
  clessnize::theme_clean_light() +
  theme(axis.text.y = element_text(size = 10)) +  # Texte plus grand
  # Étiquettes des axes et titre
  xlab("Score z") +
  ylab(NULL) +
  # Utilisation des couleurs pour les points
  scale_color_identity() +
    ggtitle(paste("Scores z des variables pour le cluster", cluster_id))
  
  # Enregistrer le graphique
  output_file <- file.path(output_dir, paste0("2021_6cluster_", cluster_id, ".png"))
  ggsave(filename = output_file, plot = graph, width = 8, height = 6)
}


generate_persona_prompt <- function(df_all_clusters, cluster_id, z_threshold = 0.25) {
  # Filtrage des données pour le cluster spécifié
  df_cluster <- df_all_clusters %>%
    filter(cluster_var == cluster_id)
  
  # Extraction des variables positives et négatives par rapport au seuil
  variables_positive <- df_cluster %>%
    filter(z_score > z_threshold) %>%
    pull(variable)
  
  variables_negative <- df_cluster %>%
    filter(z_score < -z_threshold) %>%
    pull(variable)
  
  # Construction du prompt
  prompt <- paste0(
    "\nPour le persona ", cluster_id, " :\n",
    "\nLes caractéristiques suivantes définissent ce cluster par leur distinction marquée :\n\n",
    "Variables supérieures (valeurs élevées) :\n",
    if (length(variables_positive) > 0) {
      paste(variables_positive, collapse = "\n")
    } else {
      "Aucune variable supérieure ne se distingue fortement."
    },
    "\n\nVariables inférieures (valeurs faibles) :\n",
    if (length(variables_negative) > 0) {
      paste(variables_negative, collapse = "\n")
    } else {
      "Aucune variable inférieure ne se distingue fortement."
    },
    "\n\nSuggérez un nom qui reflète bien ces caractéristiques.\n",
    "Tu peux donner un prénom significatif suivi par un deux-points (:) et un nom descriptif.\n",
    "Finalement, décris la persona en 2-3 phrases, en tenant compte des scores z des variables pour ce cluster.\n\n"
  )
  
  return(prompt)
}

# Exemple d'utilisation avec df_cluster (remplacer par votre dataframe réel)
prompt_persona <- generate_persona_prompt(df_mean_by_cluster, cluster_id = 6)
cat(prompt_persona)

# Cluster par categories
data_filtered$cluster_labels <- NA
data_filtered$cluster_labels[data_filtered$cluster_6 == 1] <- "1. Sophia - Non-franco Urbaine"
data_filtered$cluster_labels[data_filtered$cluster_6 == 2] <- "2. Stéphane - Travailleur Discret"
data_filtered$cluster_labels[data_filtered$cluster_6 == 3] <- "3. Éloïse - Jeune Sportive"
data_filtered$cluster_labels[data_filtered$cluster_6 == 4] <- "4. Gabriel - L'Urbain Raffiné"
data_filtered$cluster_labels[data_filtered$cluster_6 == 5] <- "5. Zoé - Écolo Avant-gardiste"
data_filtered$cluster_labels[data_filtered$cluster_6 == 6] <- "6. Michel - Senior Traditionnaliste"


#Nom du persona 1: Sophia : Non-franco urbaine
#35 à 54 ans, est une personne urbaine et active qui adopte un style de vie équilibré.
#Utilisatrice régulière des transports en commun, elle pratique le yoga et préfère le café Starbucks
#pour ses moments de détente. Immigrante parlant une autre langue que le français ou l’anglais,
#elle se distingue par son choix de ne jamais avoir fumé et par des habitudes qui reflètent un intérêt
#pour le bien-être et la simplicité citadine.
#est une anglophone montréalaise typique, préférant la vie en milieu urbain
#et évitant les banlieues avec leurs maisons individuelles. 

#Nom du persona 2: Stéphane :  Travailleur Discret
#Description :



#Nom du persona 3: Éloïse : Jeune Sportive
#Description :
#Le groupe d’Éloïse se distingue par des valeurs élevées pour des activités comme le gym,
#les sports d’équipe, la course et les activités motorisées. Ils sont majoritairement âgés
#de moins de 34 ans, n’ont jamais fumé et préfèrent les cocktails. Francophones,
#ils affichent des valeurs faibles pour les catégories liées à l’âge (35 ans et plus),
#aux grandes surfaces, au café, au vin rouge, et sont moins susceptibles d’être des hommes
#ou d’avoir arrêté de fumer.

#Nom du persona 4: Gabriel : L'Urbain Raffiné
#Description du persona :
#Gabriel est un homme gai urbain et sophistiqué, vivant dans un condo et incarnant un style de vie raffiné.
#Ce cluster se démarque par un fort intérêt pour les activités culturelles comme les musées,
#un goût prononcé pour le vin rouge, et une préférence pour le transport public,
#reflétant une sensibilité esthétique et un engagement envers un mode de vie urbain.

#Nom du persona 5: Zoé : L'écolo avant-gardiste 
#Description :
#Zoé incarne une génération jeune et active, très engagée dans des habitudes de vie saines
#et respectueuses de l'environnement.
#Elle privilégie des activités comme le yoga, les visites culturelles et la marche comme mode de transport,
#tout en adoptant un régime alimentaire végétarien ou végan.
#Préférant des choix modernes comme le café Starbucks,
#elle se distingue aussi par son ouverture d'esprit, notamment sur le plan de l'orientation sexuelle.


#Nom du persona 6: Michel : Le Senior Traditionnaliste

#Description du persona :
#Michel incarne les valeurs traditionnelles d'une génération mature et stable.
#Ce persona représente principalement des individus âgés de 55 ans ou plus, ayant fait des choix de vie
#marqués par la simplicité et la modération : ils ont arrêté de fumer, ne consomment pas de café
#et affichent un mode de vie sobre. Leur identité francophone et hétérosexuelle est un aspect central,
#tandis qu’ils se distinguent par une faible propension à adopter des comportements modernes ou audacieux,
#comme le tatouage ou la course à pied.
